KMS-532: Move functions to getCsvPaths
