--- serverless/src/utils/fetchNarrowers.js
@@ -1,17 +1,0 @@
-const fetchNarrowers = (uri, map) => {
-  const triples = map[uri] || []
-
-  const results = triples.map((item) => {
-    const { prefLabel, narrower, narrowerPrefLabel } = item
-
-    return {
-      prefLabel: prefLabel.value,
-      narrowerPrefLabel: narrowerPrefLabel.value,
-      uri: narrower.value
-    }
-  })
-
-  return results
-}
-
-export default fetchNarrowers
